Use python-2 official docker image to fix build issue
The build was failing because the pypi version could not find
Stripe v1.31 to install.  Not sure why this suddenly stopped building.

This commit changes the base image to use the official Python 2
image, giving us some helpful abstraction so we don't have to manage
the python installation or Ubuntu upgrades.