Install newrelic python agent; wrap heroku web & celery startup commands
This commit installs the newrelic python agent (using pip), and
wraps both the web and celery commands with the newrelic agent.

This follows the newrelic python agent documentation here:

- "Install the Python Agent": https://docs.newrelic.com/docs/apm/agents/python-agent/hosting-services/python-agent-heroku/#installing-the-python-agent
- "Track Celery Tasks": https://docs.newrelic.com/docs/apm/agents/python-agent/hosting-services/python-agent-heroku/#tracking-of-celery-tasks

We have not yet installed the Heroku Newrelic add-on, which may be necessary.
Trying to see if we can configure the app to send data directly to our
existing newrelic account, rather than having the add-on create a newrelic
account that's only accessible through the Heroku interface.