Remove Newrelic
This commit removes New Relic from the heroku app.

We are doing this because the Papertrail logs are being cluttered with warnings from the newrelic collector.  Since we rarely, if ever, use newrelic on this app, we are removing the resource add-on for now.  We will reinstall it if we find that it would be helpful.

Example WARNING log:
Mar 21 16:58:10 donations-prod app/web.1 2023-03-21 16:58:10,341 (51/NR-Harvest-Thread) newrelic.core.agent_protocol WARNING - Received a non 200 or 202 HTTP response from the data collector where params={..., status_code=403 and content=b'{}'